requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
openai==0.27.0
discord.py==2.2.2
requests==2.28.2
beautifulsoup4==4.11.2
ccxt==2.9.6